CONTAINS BUG: Refactoring of Application Dropdown (Bug: not passing the taskList from Dropdown to parent)
